log + progress bar finished
log + progress bar finished
need to refresh list on 3-rd tab on tab-select

+++ CompareMDBs/CompareMDBs.cs
@@ -87,22 +87,26 @@ namespace CompareMDBs
                     //MessageBox.Show("All tables have been inserted!");
                     //MessageBox.Show("Все таблицы были импортированы!");
                     lbl_conditionTab1.Text = "Все таблицы были импортированы! Идёт подготовка БД.";
+                    lbl_conditionTab1.Refresh();
                     progressBar1.Value += 3;
 
                     Program.prepareDB(".\\Teamsoft.mdb");
                     progressBar1.Value += 10;
                     lbl_conditionTab1.Text = "БД подготовлена. Устанавливаются первичные ключи.";
+                    lbl_conditionTab1.Refresh();
                     Program.setPrimaryKeys(".\\Teamsoft.mdb");
                     progressBar1.Value += 35;
 
                     try
                     {
                         lbl_conditionTab1.Text = "Первичные ключи установлены. Идёт сжатие БД, поождите.";
+                        lbl_conditionTab1.Refresh();
                         Program.compactAndRepair(".\\Teamsoft.mdb");
                         progressBar1.Value += 50;
                         //MessageBox.Show("Database compacted!");
                         //MessageBox.Show("Сжатие базы данных выполнено.");
                         lbl_conditionTab1.Text = "Сжатие базы данных выполнено.";
+                        lbl_conditionTab1.Refresh();
                     }
                     catch (Exception e2)
                     {
@@ -129,11 +133,22 @@ namespace CompareMDBs
                 else
                 {
                     if (Program.GetDBTables(".\\Teamsoft.mdb").Contains(txtDelTable.Text.ToString()))
+                    {
                         Program.dropThisShit(".\\Teamsoft.mdb", txtDelTable.Text.ToString(), true);
+                        /*
+                        if (lst_allTables.Items.Count > 0 && lst_allTables.Items.Contains(txtDelTable.Text.ToString()))
+                        {
+                            lst_allTables.Items.Remove(txtDelTable.Text.ToString());
+                            lst_allTables.Refresh();
+                        }
+                        */
+                    }
+                    else if (txtDelTable.Text.ToString().Length == 0)
+                        MessageBox.Show("Название таблицы не введено!");
                     else //MessageBox.Show("Table with name " + txtDelTable.Text.ToString() +
-                         //               " couldn't be found in database. Try drop another table.");
-                         MessageBox.Show("Таблицу с названием " + txtDelTable.Text.ToString() +
-                                   " не удалось найти в базе данных. Проверьте, правильно ли ввели название таблицы.");
+                        //               " couldn't be found in database. Try drop another table.");
+                        MessageBox.Show("Таблицу с названием '" + txtDelTable.Text.ToString() +
+                                  "' не удалось найти в базе данных. Проверьте, правильно ли ввели название таблицы.");
                 }
             }
             catch (OleDbException ex)
@@ -146,38 +161,63 @@ namespace CompareMDBs
         //delete old data
         private void btn_deleteOldData_Click(object sender, EventArgs e)
         {
+            progressBar3.Value = 0;
+            lbl_conditionTab3.Text = "";
+            lbl_conditionTab3.Refresh();
+
             string conString = ("Provider=Microsoft.JET.OLEDB.4.0;data source= .\\teamsoft.mdb;Persist Security Info=False;");
             OleDbConnection dbconn = new OleDbConnection(conString);
             dbconn.Open();
             OleDbCommand dbcommand = new OleDbCommand();
 
-            foreach (Object item in lst_allTables.CheckedItems)
+            if (lst_allTables.CheckedItems.Count > 0)
             {
-                DataTable columns = dbconn.GetOleDbSchemaTable(OleDbSchemaGuid.Columns,
-                                                        new object[] { null, null, item.ToString(), null });
-                string delQuery = "";
-                var schema = dbconn.GetSchema("COLUMNS");
-                string dueDate = pckr_dueDate.Value.ToString("yyyy-MM-dd");
-                //choose values in what column should be compared with selected date
-                if (schema.Select("TABLE_NAME='" + item.ToString() + "' AND COLUMN_NAME='datefrom'").Length > 0)
-                    delQuery = "DELETE FROM " + item.ToString() + " where datefrom < #" + dueDate + "#";
-                else if (schema.Select("TABLE_NAME='" + item.ToString() + "' AND COLUMN_NAME='date'").Length > 0)
-                    delQuery = "DELETE FROM " + item.ToString() + " where date < #" + dueDate + "#";
-                else
-                    delQuery = "DELETE FROM " + item.ToString() + " where currenttime < #" + dueDate + "#";
-                
-                dbcommand.CommandText = delQuery;
-                dbcommand.CommandType = CommandType.Text;
-                dbcommand.Connection = dbconn;
+                progressBar3.Maximum = lst_allTables.CheckedItems.Count;// +10;
+                foreach (Object item in lst_allTables.CheckedItems)
+                {
+                    //System.Threading.Thread.Sleep(1000);
+                    lbl_conditionTab3.Text = "Удаляются данные из таблицы " + item.ToString();
+                    lbl_conditionTab3.Refresh();
+                    //System.Threading.Thread.Sleep(1000);
+                    DataTable columns = dbconn.GetOleDbSchemaTable(OleDbSchemaGuid.Columns,
+                                                            new object[] { null, null, item.ToString(), null });
+                    string delQuery = "";
+                    var schema = dbconn.GetSchema("COLUMNS");
+                    string dueDate = pckr_dueDate.Value.ToString("yyyy-MM-dd");
+                    //choose values in what column should be compared with selected date
+                    if (schema.Select("TABLE_NAME='" + item.ToString() + "' AND COLUMN_NAME='datefrom'").Length > 0)
+                        delQuery = "DELETE FROM " + item.ToString() + " where datefrom < #" + dueDate + "#";
+                    else if (schema.Select("TABLE_NAME='" + item.ToString() + "' AND COLUMN_NAME='date'").Length > 0)
+                        delQuery = "DELETE FROM " + item.ToString() + " where date < #" + dueDate + "#";
+                    else
+                        delQuery = "DELETE FROM " + item.ToString() + " where currenttime < #" + dueDate + "#";
 
-                int result = Program.ConvertFromDBVal<int>(dbcommand.ExecuteNonQuery()); 
-                //MessageBox.Show("Old values from table '" + item.ToString() + "' were deleted!");
-                Program.addToLog("Записи, внесённые ранее " + dueDate.ToString() + ", были удалены из таблицы " + item.ToString());
-            }
+                    dbcommand.CommandText = delQuery;
+                    dbcommand.CommandType = CommandType.Text;
+                    dbcommand.Connection = dbconn;
+
+                    int result = Program.ConvertFromDBVal<int>(dbcommand.ExecuteNonQuery());
+                    //MessageBox.Show("Old values from table '" + item.ToString() + "' were deleted!");
+                    Program.addToLog("Записи, внесённые ранее " + dueDate.ToString() + ", были удалены из таблицы " + item.ToString());
+                    progressBar3.Value += 1;
+                }
             dbconn.Close();
+            /*
+            lbl_conditionTab3.Text = "Старые данные из выбранных таблиц удалены. Выполняется сжатие БД.";
+            lbl_conditionTab3.Refresh();
             Program.compactAndRepair(".\\Teamsoft.mdb");
+            progressBar3.Value += 7;
+            lbl_conditionTab3.Text = "Сжатие БД завершено. Выполняется проставление недостающих индексов.";
+            lbl_conditionTab3.Refresh();
+            Program.CreateIndexes(".\\Teamsoft.mdb");
+            progressBar3.Value += 3;
+            */
+            lbl_conditionTab3.Text = "Старые данные из выбранных таблиц удалены!";
+            lbl_conditionTab3.Refresh();
             //MessageBox.Show("Old values from all selected tables were deleted!");
-            MessageBox.Show("Старые данные из выбранных таблиц были удалены!");
+            //MessageBox.Show("Старые данные из выбранных таблиц были удалены!");
+            }
+            else MessageBox.Show("Хоть одну таблицу из списка выбрать всё-таки нужно.");
         }
         
         //export tables checked in listview to new MDB
@@ -213,29 +253,119 @@ namespace CompareMDBs
             {
                 progressBar1.Value = 0;
                 lbl_conditionTab1.Text = "";
+                lbl_conditionTab1.Refresh();
+
                 if (!File.Exists(".\\Teamsoft.mdb"))
                     //MessageBox.Show("File 'Teamsoft.mdb' doesn't exist in current directory.");
                     MessageBox.Show("Файл 'Teamsoft.mdb' не существует в текущей папке.");
                 else
                 {
                     progressBar1.Value += 5;
-                    
+
                     lbl_conditionTab1.Text = "Идёт подготовка БД.";
+                    lbl_conditionTab1.Refresh();
                     Program.prepareDB(".\\Teamsoft.mdb");
                     progressBar1.Value += 25;
-                    
+
                     lbl_conditionTab1.Text = "БД подготовлена. Идёт сжатие БД, поождите.";
+                    lbl_conditionTab1.Refresh();
                     Program.compactAndRepair(".\\Teamsoft.mdb");
                     progressBar1.Value += 50;
-                    
+
                     lbl_conditionTab1.Text = "Сжатие базы данных выполнено. Проставляем индексы.";
+                    lbl_conditionTab1.Refresh();
                     Program.CreateIndexes(".\\Teamsoft.mdb");
                     progressBar1.Value += 20;
                     //MessageBox.Show("Сжатие базы данных выполнено, индексирование проведено.");
                     lbl_conditionTab1.Text = "Сжатие базы данных выполнено, индексирование проведено.";
+                    lbl_conditionTab1.Refresh();
                 }
             }
             catch(OleDbException ex)
+            {
+                //MessageBox.Show("Oops! \n" + ex.Message + " \n Try compare MDB with MS Access");
+                MessageBox.Show("Упс! \n" + ex.Message + " \n Попробуйте сжать/востановить базу в MS Access");
+            }
+        }
+
+        private void btn_compact3_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                progressBar3.Value = 0;
+                progressBar3.Maximum = 100;
+                lbl_conditionTab3.Text = "";
+                lbl_conditionTab3.Refresh();
+
+                if (!File.Exists(".\\Teamsoft.mdb"))
+                    //MessageBox.Show("File 'Teamsoft.mdb' doesn't exist in current directory.");
+                    MessageBox.Show("Файл 'Teamsoft.mdb' не существует в текущей папке.");
+                else
+                {
+                    progressBar3.Value += 5;
+
+                    lbl_conditionTab3.Text = "Идёт подготовка БД.";
+                    lbl_conditionTab3.Refresh();
+                    Program.prepareDB(".\\Teamsoft.mdb");
+                    progressBar3.Value += 25;
+
+                    lbl_conditionTab3.Text = "БД подготовлена. Идёт сжатие БД, поождите.";
+                    lbl_conditionTab3.Refresh();
+                    Program.compactAndRepair(".\\Teamsoft.mdb");
+                    progressBar3.Value += 50;
+
+                    lbl_conditionTab3.Text = "Сжатие базы данных выполнено. Проставляем индексы.";
+                    lbl_conditionTab3.Refresh();
+                    Program.CreateIndexes(".\\Teamsoft.mdb");
+                    progressBar3.Value += 20;
+                    //MessageBox.Show("Сжатие базы данных выполнено, индексирование проведено.");
+                    lbl_conditionTab3.Text = "Сжатие базы данных выполнено, индексирование проведено.";
+                    lbl_conditionTab3.Refresh();
+                }
+            }
+            catch (OleDbException ex)
+            {
+                //MessageBox.Show("Oops! \n" + ex.Message + " \n Try compare MDB with MS Access");
+                MessageBox.Show("Упс! \n" + ex.Message + " \n Попробуйте сжать/востановить базу в MS Access");
+            }
+        }
+
+        private void btn_compact2_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                progressBar2.Value = 0;
+                progressBar2.Maximum = 100;
+                lbl_conditionTab2.Text = "";
+                lbl_conditionTab2.Refresh();
+
+                if (!File.Exists(".\\Teamsoft.mdb"))
+                    //MessageBox.Show("File 'Teamsoft.mdb' doesn't exist in current directory.");
+                    MessageBox.Show("Файл 'Teamsoft.mdb' не существует в текущей папке.");
+                else
+                {
+                    progressBar2.Value += 5;
+
+                    lbl_conditionTab2.Text = "Идёт подготовка БД.";
+                    lbl_conditionTab2.Refresh();
+                    Program.prepareDB(".\\Teamsoft.mdb");
+                    progressBar2.Value += 25;
+
+                    lbl_conditionTab2.Text = "БД подготовлена. Идёт сжатие БД, поождите.";
+                    lbl_conditionTab2.Refresh();
+                    Program.compactAndRepair(".\\Teamsoft.mdb");
+                    progressBar2.Value += 50;
+
+                    lbl_conditionTab2.Text = "Сжатие базы данных выполнено. Проставляем индексы.";
+                    lbl_conditionTab2.Refresh();
+                    Program.CreateIndexes(".\\Teamsoft.mdb");
+                    progressBar2.Value += 20;
+                    //MessageBox.Show("Сжатие базы данных выполнено, индексирование проведено.");
+                    lbl_conditionTab2.Text = "Сжатие базы данных выполнено, индексирование проведено.";
+                    lbl_conditionTab2.Refresh();
+                }
+            }
+            catch (OleDbException ex)
             {
                 //MessageBox.Show("Oops! \n" + ex.Message + " \n Try compare MDB with MS Access");
                 MessageBox.Show("Упс! \n" + ex.Message + " \n Попробуйте сжать/востановить базу в MS Access");
